sequence tile
