+++ common/src/main/webapp/sass/explore.scss
@@ -1323,12 +1323,13 @@ $search-preview-margin: 5px;
 
 .explore-sequence-tile-title {
   display: flex;
+  justify-content: center;
   gap: 1rem;
 
   // Hide the info icon on small screens
-  > :first-child {
-    display: none;
-  }
+  // > :first-child {
+  //   display: none;
+  // }
 
   .explore-sequence-tile-title-item {
     font-weight: bold;
